+++ backend/tests/test_utils.py
@@ -87,26 +87,6 @@ class TestUtils(TestCase):
         # test fallback
         self.assertEqual(util.parsedatetime('blyf', 'flaf'), 'flaf')
 
-    def test_splitlist(self):
-        self.assertEqual(
-            list(util.splitlist([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], 3)),
-            [[1, 2, 3], [4, 5, 6], [7, 8, 9], [10]],
-        )
-        self.assertEqual(
-            list(util.splitlist([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], 4)),
-            [[1, 2, 3, 4], [5, 6, 7, 8], [9, 10]],
-        )
-        self.assertEqual(
-            list(util.splitlist([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], 11)),
-            [[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]],
-        )
-        self.assertRaises(exceptions.HTTPException,
-                          list, util.splitlist([], 0))
-        self.assertRaises(exceptions.HTTPException,
-                          list, util.splitlist([], -1))
-        self.assertRaises(TypeError,
-                          list, util.splitlist([], 'horse'))
-
     def test_is_uuid(self):
         self.assertTrue(util.is_uuid('00000000-0000-0000-0000-000000000000'))
         self.assertFalse(util.is_uuid('42'))